Issue #50
Changed line 10 on page.tsx in the officers folder to default to text center for the first paragraph and on large screens to text-left for the header and same thing on line 11 for the paragraph.

 Changed line 25 on Department.tsx on the officers folder. Added a default left margin for 5rem and added a conditional for large screens to have a left margin of 0.

Changed People.tsx to fix margins as well. Changed line 22 on the page to have a default left margin of 3.75rem and on large screens to have a left margin of 0.